Scopes: remove the URL sync manager and rely on the facade to update the URL values
